Revert "Do not send watchdog while sending or receiving"
The idea is:
If the TLC is busy sending many messages, for instance sending
buffered messages directly after connecting establishment. In that
case the TLC might not respond to Watchdog messages quickly enough to
Watchdog messages, and thus forcing the supervisor to disconnect.

But this problem is already covered with
c6190f85e1bec18cce760040db922aef68eed7a3

It still a good idea to limit watchdog messages as discussed in #47,
but we're reverting this for now since it's a breaking change

This reverts e90a77a85d1226c3cc6c1e06b86330188c72ae7d