Support subscriptions with update on change+interval
Adds sendOnChange "sOc" which can be used in parallel with uRt.

We've observed that sometimes the supervision system doesn't show the
recent status of the equipment correctly. Due to the fact that in many
cases StatusSubscribes only update on change - and some statuses changes
very seldomly - may cause the supervision system to show incorrect or
missing status for an equipment during a long time.

Previously, StatusSubscribe supported either update on change (uRt=0) or
interval (uRt>0).

Discussed in #21