Support subscriptions with update on change+interval
Adjust meaning of updateRate (uRt) when set to "0".

When set to "0" it means that interval updates are turned off since
otherwise sendOnChange (sOc) can be set to "true" at the same time
updateRate is set to "0" which is conflicting. There is also a
need to turn interval updates off in case only updateOnChange is
requested.